+++ app/models/user_event.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+# Tracks notification emails for user events.
+class UserEvent < ApplicationRecord
+  # Relationships
+  belongs_to :user
+  belongs_to :event
+end