+++ db/migrate/20171113204045_remove_brain_surveys_count_from_users.rb
@@ -1,0 +1,5 @@
+class RemoveBrainSurveysCountFromUsers < ActiveRecord::Migration[5.1]
+  def change
+    remove_column :users, :brain_surveys_count, :integer, null: false, default: 0
+  end
+end