+++ test/controllers/internal_controller_test.rb
@@ -82,7 +82,7 @@ class InternalControllerTest < ActionDispatch::IntegrationTest
     post test_my_brain_start_url
     @consented.reload
     assert_not_nil @consented.brain_started_at
-    assert_redirected_to "#{ENV["test_my_brain_url"]}?id=#{@consented.subject_code}#{@consented.current_event.upcase}"
+    assert_redirected_to "#{ENV["test_my_brain_url"]}?id=#{@consented.subject_code}#{@consented.current_event_brain_code}"
   end
 
   test "should complete test my brain for consented user" do